added wishlist feature

+++ app/src/main/java/kalbe/corp/genexsupabasepoc/models/Product.kt
@@ -6,10 +6,10 @@ import kotlinx.serialization.Serializable
 data class Product(
     val id: String,
     val name: String,
-    val price: Float,
+    val price: Double,
     val image: String,
     val capsules: Int,
-    val rating: Float,
+    val rating: Double,
     val description: String,
     val category: String,
 )
